Move Client implementation from test to Client folder
