+++ packages/cpp/ArmoniK.Api.Client/source/submitter/SessionContext.cpp
